greybus: i2c: fix error handling
The i2c protocol is sending the bad error number when the i2c device
is not responding. In thekernel side, this was resulting with an error:
"transfer operation failed (-5)".

Testing Done: transfer operation failed (-5) is not longer printed
Bug Number: Trivial
Review URL:
Reviewed by:

+++ nuttx/drivers/greybus/i2c.c
@@ -65,7 +65,7 @@ static uint8_t gb_i2c_protocol_transfer(struct gb_operation *operation)
 {
     int i, op_count;
     uint32_t size = 0;
-    uint8_t ret;
+    int ret;
     uint8_t *write_data;
     bool read_op;
     int read_count = 0;
